requirements.txt
streamlit==1.31.0
openai==0.28.0
python-dotenv==1.0.1
google-generativeai==0.3.2
